pyproject.toml
[tool.poetry]
name = "vim"
version = "0.1.0"
description = ""
authors = ["DDMAL Lab"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.5"
psycopg = "3.1.10"
requests = "2.31.0"
gunicorn = "21.2.0"
pillow = "^10.4.0"
django-vite = "^3.1.0"
pysolr = "^3.10.0"
django-cleanup = "^8.0.0"
django-ratelimit = "^4.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
mypy = "^1.5.1"
pylint = "^2.17.6"
types-requests = "^2.32.0.20240712"
django-stubs = "^5.0.4"
djlint = "^1.34.1"


[tool.poetry.group.debug.dependencies]
django-debug-toolbar = "^4.2.0"
django-extensions = "3.2.3"
werkzeug = "3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
